requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.19
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.5
psycopg-binary==3.2.5
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
